requirements.txt
# Combined requirements from agent_marketplace and mcp_router
# AI/ML and Data Science packages
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
asyncio==3.4.3
attrs==25.3.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.7.0
build==1.2.2.post1

# Web Framework packages
Flask==3.0.0
Flask-CORS==4.0.0
fastapi==0.115.12
fastmcp
uvicorn==0.34.3
starlette==0.46.2

# Cache and Storage
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1

# Security and Crypto
cryptography==45.0.4
PyJWT==2.10.1
python_bcrypt==0.3.2
pyOpenSSL==25.1.0

# Data and JSON handling
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
docopt==0.6.2
durationpy==0.10

# Environment and Config
python-dotenv==1.1.0

# Excel and File handling
et-xmlfile==2.0.0
executing==2.2.0
faiss-cpu==1.11.0
fastjsonschema==2.21.1
filelock==3.18.0

# Flask extensions
# Removed duplicate flask entries - using Flask==3.0.0 and Flask-CORS==4.0.0 from above
flask-restx==1.3.0
flask-sqlalchemy==3.1.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1

# Web server and networking
gevent==23.9.1
google==3.0.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gotrue==2.12.3
greenlet==3.2.3
grpcio==1.73.0
gunicorn==21.2.0

# HTTP libraries
h11==0.16.0
h2==4.2.0
hf-xet==1.1.3
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0

# AI/ML libraries
huggingface-hub==0.33.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
iniconfig==2.1.0

# Jupyter and IPython
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1

# JSON processing
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyterlab-pygments==0.3.0

# Kubernetes and cloud
kubernetes==33.1.0

# LangChain ecosystem
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-experimental==0.3.4
langchain-huggingface==0.3.0
langchain-mcp-adapters==0.1.7
langchain-openai==0.3.23
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45

# AI/ML core libraries
litellm==1.72.6
lxml==4.9.4
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7

# MCP (Model Context Protocol)
mcp==1.9.4
mcpadapt==0.1.10
mdurl==0.1.2
mistune==3.1.3
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.4
mypy-extensions==1.1.0

# Jupyter notebook handling
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.5
nltk==3.9.1
numpy==2.3.0

# NVIDIA CUDA libraries
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77

# OAuth and authentication
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.86.0
openpyxl==3.1.2

# OpenTelemetry (observability)
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1

# Data serialization
orjson==3.10.18
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pipreqs==0.5.0
platformdirs==4.3.8

# Browser automation
playwright==1.52.0
pluggy==1.6.0

# Supabase and database
postgrest==1.1.1
posthog==4.10.0
prompt-toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
ptyprocess==0.7.0
pure-eval==0.2.3

# Authentication and security
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.9.1
pyee==13.0.0
Pygments==2.19.1
pyparsing==3.2.3
pypika==0.48.9
pyproject-hooks==1.2.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.14.1

# Date and time handling
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.0

# Search and ranking
rank-bm25==0.2.2
realtime==2.6.0
referencing==0.36.2
regex==2024.11.6

# HTTP requests
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0

# Terminal and output formatting
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1

# ML model handling
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smolagents==1.18.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
sse-starlette==2.3.6
stack-data==0.6.3

# Database and storage
storage3==0.12.0
strenum==0.4.15
supabase==2.17.0
supafunc==0.10.1
sympy==1.14.0

# Utilities and tools
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1

# PyTorch and deep learning
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
triton==3.3.1

# CLI and typing
typer==0.16.0
typing-extensions==4.14.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0

# ASGI and async
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1

# WebSocket and networking
websocket-client==1.8.0
websockets==15.0.1
werkzeug==3.0.1
wrapt==1.17.2

# Hashing and compression
xxhash==3.5.0
yarg==0.1.9
yarl==1.20.1
zipp==3.23.0
zope-event==5.1
zope-interface==7.2
zstandard==0.23.0

# Legacy packages from agent_marketplace (smaller footprint)
Brotli==1.1.0
brotlicffi==1.1.0.0
curio==1.6
cython==3.1.2
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
hypothesis==6.135.26
outcome==1.3.0.post0
python_socks==2.7.1
simplejson==3.20.1
socksio==1.0.0
speedups==1.4.0
thread==2.0.5
toml==0.10.2
tomli==2.2.1
watchdog==6.0.0